reinforcement_learning/requirements.txt
aiofiles==0.8.0
appdirs==1.4.4
autopep8==1.6.0
bowler==0.9.0
box2d-py==2.3.5
certifi==2022.9.24
charset-normalizer==2.1.1
cloudpickle==2.2.0
docker==5.0.3
dockerpty==0.4.1
fissix==21.11.13
gym==0.26.2
gym-notices==0.0.8
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
moreorless==0.4.0
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
pygame==2.1.0
python-dateutil==2.8.2
python-gnupg==0.4.9
pytz==2022.6
requests==2.28.1
six==1.16.0
swig==4.1.0
typing_extensions==4.4.0
urllib3==1.26.12
volatile==2.1.0
webencodings==0.5.1
websocket-client==1.4.2
yapf==0.32.0
zipp==3.10.0
mdptoolbox-hiive==4.0.3.1